Add files via upload
This project focuses on analyzing the Crime and Disorder Data from the City's data portal to forecast future crime occurrences. The dataset spans from 2018 to 2024 and records the monthly crime counts. After conducting an in-depth analysis of the data, I will develop and optimize a neural network model to predict future crime trends.